Use MARISA_DEBUG_IF in scoped-*.h as opposed to MARISA_THROW_IF
These are the only two places that we throw exceptions from header files
in non-debug mode.

With this change, developers can take the marisa library without being
forced to enable exceptions for their whole library / application.

IMO, self-reset does not seem to be fatal enough to throw in production
and MARISA_DEBUG_IF should be enough for us to catch it if we think it
should never happen.