Add and use .clang-format (#82)
Reformat all files with
```
clang-format -i lib/**/*.{h,cc} include/**/*.h
```
.

Some of the existing formatting conventions could not be replicated with
clang-format (see the diffs).  The most notable is the alignment of enum
definitions.  marisa code likes these to be aligned, but clang-format
only has an `AlignConsecutiveAssignments` option that controls all
assignments.