Use `uint32_t` instead of `size_t` where possible
Marisa node IDs are 32-bit integers, so using a 64-bit `size_t`
for nodes and related types wastes memory.
Additionally, key and query lengths etc are all effectively
limited to 32-bit.

Replaces `size_t` with `uint32_t` where appropriate.
This reduces the sizes of various structs throughout.

Also replaces the remaining `marisa::UInt*` usages with `<cstdint>`
types. I've opted to not use the `std::` prefix for `<cstdint>`
types because it is verbose and not required.